% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity_Bede.R
\name{Diversity_Bede}
\alias{Diversity_Bede}
\title{Model Function}
\usage{
Diversity_Bede(
  Input_df,
  Meta = 1:3,
  Total.Abundance = TRUE,
  Number.Taxa = TRUE,
  Shannons = TRUE,
  Pielou = TRUE,
  Simpsons = TRUE,
  Rarefaction = TRUE,
  NRare = 50,
  output = "Long"
)
}
\arguments{
\item{Input_df}{a wide dataframe of Species by Site.}

\item{Total.Abundance}{logical: if you want this metric. Default is TRUE}

\item{Number.Taxa}{logical: if you want this metric. Default is TRUE}

\item{Shannons}{logical: if you want this metric. Default is TRUE}

\item{Pielou}{logical: if you want this metric. Default is TRUE}

\item{Simpsons}{logical: if you want this metric. Default is TRUE}

\item{NRare}{number of species for Rarefy (see vegan). Default is 50}

\item{output}{output "Long" or "Wide". "Long" is default}

\item{Rare}{logical: if you want this metric (See vegan). Default is TRUE}
}
\description{
This Function applies a glm/glmm models applying model selection using AICc to input into plot_model_Bede()
It can handle beta, gamma and poisson models with ease, gaussian is currently in development
}
\examples{
SpeciesList<-paste0("Species",seq_along(1:15)) # 15 fake species

SiteList<-paste0("Site",seq_along(1:50)) # 15 fake sites
SitexSpeciesdf<- expand_grid(Site=SiteList,Species=SpeciesList) \%>\% # create df of all unique combinations of these lists
dplyr::mutate(Abundance=rpois(n(),1.2)) \%>\% ## add fake abundance (can be count like here or continuous)
tidyr::pivot_wider(names_from = "Species",values_from = "Abundance") \%>\%
dplyr::mutate(Meta1=rep(c("Yay","Nay"),0.5*nrow(.)),
Meta2=rep(c("Nay","Yay"),0.5*nrow(.)),
.before=1)

Metrics<-Diversity_Bede(SitexSpeciesdf,Meta=1:3,
                        Total.Abundance=TRUE,
                        Number.Taxa=TRUE,
                        Shannons=TRUE,
                        Pielou=TRUE,
                        Simpsons=TRUE,
                        Rare=TRUE,
                        NRare=50,
                        output="Long")
}
\keyword{Model}
