% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_signif.R
\name{down_signif}
\alias{down_signif}
\title{Signif Down}
\usage{
down_signif(x, digits = 0)
}
\arguments{
\item{x}{data that you want to remove rows containing NAs in.}

\item{digits}{number of significant digits to change values to}
}
\description{
This function applies signif figure adjustment but rounds down
}
\examples{

down_signif(3599, digits = 2)
3500
down_signif(7890349, digits = 2)
7800000


}
\keyword{Round}
\keyword{down}
\keyword{figures}
\keyword{signif}
\keyword{to}
