% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertRow.R
\name{Insert_Row}
\alias{Insert_Row}
\title{Insert Row}
\usage{
Insert_Row(existingDF, newrow, r)
}
\arguments{
\item{Insert}{rows}
}
\description{
This function adds a row to a df
}
\examples{
Insert_Row()
}
\keyword{Insert}
