% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertRow.R
\name{insertRow}
\alias{insertRow}
\title{Insert Row}
\usage{
insertRow(data, newrow, r)
}
\arguments{
\item{data}{the input dataframe}

\item{newrow}{a vector be added as a row, must be the same length as the number of coloumns in data.}

\item{r}{the row index where the new row will be added into the dataframe.}
}
\description{
This function adds a row to a df
}
\examples{
df<-data.frame(a=c(1,3,4,5),b=c("a","c","d","e"),c=c("z","x","w","u"))
df_1<-insertRow(df,c(2,"b","y"),2)
df_1
}
\keyword{Insert}
