% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Check.R
\name{Model_Check}
\alias{Model_Check}
\title{A Model Diagnostic Function Plot}
\usage{
Model_Check(Model)
}
\arguments{
\item{Model}{a model output object.}
}
\description{
This function Plots Model Diagnostics fitted values vs residuals, qqnorm plot, histogram of residuals and Auto Correlation Factor.
}
\examples{

model <- lm(mpg ~ wt + cyl, data = mtcars)

Model_Check(model)
}
\keyword{Diagnostics}
\keyword{Model}
