% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_Bede.R
\name{plot_model_Bede}
\alias{plot_model_Bede}
\title{Model Plot Function}
\usage{
plot_model_Bede(
  df_mean,
  df_pred,
  Metric,
  Formula,
  xvar,
  colourvar,
  pallette,
  xlabel,
  Categorical.x
)
}
\arguments{
\item{df_mean}{dataframe of raw data that contains columns explicitly named Mean and CI. and two columns xvar and colourvar (named in the item xvar and colourvar)}

\item{df_pred}{dataframe of prediction data contains columns explicitly named fit_resp, upr_resp and lwr_resp. and two columns xvar and colourvar (named in the item xvar and colourvar)}

\item{Metric}{name of metric being analysed}

\item{Formula}{Formula object that will be printed above each graph}

\item{xvar}{the name of the column in df_mean and df_pred to be plotted on x axis}

\item{colourvar}{the name of the column in df_mean and df_pred to be plotted as the colour}

\item{pallette}{the colour palette to select from Utilities.Package (default="main")}

\item{xlabel}{the label to be written below x axis}

\item{Categorical.x}{logical, TRUE if x is categorical, FALSE if x is not categorical.}
}
\description{
This Function creates a ggplot of Model Predictions using a X variable (xvar) and treatment (colourvar)
}
\examples{
library(purrr)
library(patchwork)

df1<-data.frame(Response=rnorm(3000, mean=10, sd=2),
                Var1=rep(c("A","B","C"),1000),
                Var2=c(rep("X",1000),rep("Y",1000),rep("Z",1000)),
                RandomVar1=rep(c("ID_1","ID_2","ID_3","ID_4"),750),
                Metric=rep("Water_Temp",3000),
                Family=rep("gaussian",3000))

df2<-data.frame(Response=rbeta(3000,0.5,1),
                Var1=rep(c("A","B","C"),1000),
                Var2=c(rep("X",1000),rep("Y",1000),rep("Z",1000)),
                RandomVar1=rep(c("ID_1","ID_2","ID_3","ID_4"),750),
                Metric=rep("Infection_Prop",3000),
                Family=rep("beta",3000))

df3<-data.frame(Response=rgamma(3000, 30),
                Var1=rep(c("A","B","C"),1000),
                Var2=c(rep("X",1000),rep("Y",1000),rep("Z",1000)),
                RandomVar1=rep(c("ID_1","ID_2","ID_3","ID_4"),750),
                Metric=rep("Tree_Height",3000),
                Family=rep("gamma",3000))

Input_df<-rbind(df1,df2,df3) \%>\% dplyr::mutate(Var1=as.factor(Var1),Var2=as.factor(Var2),RandomVar1=as.factor(RandomVar1))

Modelled<-model_Bede(Input_df=Input_df,Response="Response",
                     Var1="Var1",Var2="Var2",RandomVar1="RandomVar1",
                     Metric="Metric",Mixed=FALSE,Family="Family")


ModelPlots <- pmap(list(df_mean=Modelled$Raw,
                            df_pred=Modelled$Prediction,
                            Metric=Modelled$Metric,
                            Formula=Modelled$Formula,
                            xvar="Var1",
                            colourvar="Var2",
                            pallette="main",
                            xlabel="Variable 1",
                            Categorical.x=TRUE),
                       plot_model_Bede)

wrap_plots(ModelPlots,ncol=2)+plot_layout(guides="collect")+plot_annotation(tag_levels = "a")

}
\keyword{Model}
\keyword{Predictions}
