% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_Bede.R
\name{scale_colour_Bede}
\alias{scale_colour_Bede}
\title{Scale Colour Function}
\usage{
scale_colour_Bede(
  palette = "main",
  alpha = 1,
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Select one of the preset colour palettes within the package. "main" is default. other options include "cool", "hot","mixed", "grey", "mussels" "bathy", "PhD", "Orkney", "Bathy_Blues" and "SAS"}

\item{discrete}{Select whether colour scale is discrete or not. Automatic unless selected. TRUE is default.}

\item{reverse}{Select which direction a scale should be applied.}
}
\description{
This function add colours to add to ggplot scales
}
\examples{

df<- data.frame(x=rep(seq(1,10,1),10),
                y=rep(rnorm(10,4),10),
                Treatment=rep(c("A","B"),50))

ggplot(df,aes(x=x,y=y,colour=Treatment))+
geom_point()+
scale_colour_Bede("SAS")

}
\keyword{Colours}
\keyword{Palettes}
