% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error Function}
\usage{
se(x)
}
\arguments{
\item{x}{vector of values to calculate Standard error from}
}
\description{
This function calculates Standard Error
}
\examples{
vec<-rnorm(200,0)
vec_se<-se(vec)
vec_se
}
\keyword{se}
