% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Bede.R
\name{theme_Bede}
\alias{theme_Bede}
\title{A ggplot Theme Function}
\usage{
theme_Bede(...)
}
\description{
This function adds a custom theme to ggplots
}
\examples{
df<- data.frame(x=rep(seq(1,10,1),10),y=rep(rnorm(10,4),10))
ggplot(df,aes(x=x,y=y))+geom_point()+theme_Bede()
}
\keyword{Theme}
