% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Bede_Map.R
\name{theme_Bede_Map}
\alias{theme_Bede_Map}
\title{A ggplot Theme Function}
\usage{
theme_Bede_Map(...)
}
\description{
This function adds a custom theme to ggplots specifcally for Maps over large reas (Region Scales)
}
\examples{
df<- data.frame(x=rep(seq(1,10,1),10),y=rep(rnorm(10,4),10))
ggplot(df,aes(x=x,y=y))+geom_point()+theme_Bede_Map()
}
\keyword{Theme}
