% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_DLCS.R
\name{theme_DLCS}
\alias{theme_DLCS}
\title{A ggplot Theme Function}
\usage{
theme_DLCS(...)
}
\description{
This function adds a custom theme to ggplots from a project on Deep Learning Convolutional Stacking
}
\examples{
df<- data.frame(x=rep(seq(1,10,1),10),y=rep(rnorm(10,4),10))
ggplot(df,aes(x=x,y=y))+geom_point()+theme_DLCS()
}
\keyword{Theme}
