% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Mareta_Dark.R
\name{theme_Mareta_Dark}
\alias{theme_Mareta_Dark}
\title{A ggplot Theme Function}
\usage{
theme_Mareta_Dark(...)
}
\description{
This function adds a custom theme to ggplots
}
\examples{
df<- data.frame(x=rep(seq(1,10,1),10),y=rep(rnorm(10,4),10))
ggplot(df,aes(x=x,y=y))+geom_point()+theme_Mareta_Dark()
}
\keyword{Theme}
