% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Bede.R
\name{model_Bede}
\alias{model_Bede}
\title{Model Function}
\usage{
model_Bede(
  Input_df,
  Response,
  Time_Element,
  Treatment_Element,
  Random_Element,
  Metric_column,
  Mixed
)
}
\arguments{
\item{Model}{glm or glmm}
}
\description{
This Function applies a glm/glmm models applying model selection using AICc to input into plot_model_Bede()
It can handle beta, gamma and poisson models with ease, gaussian is currently in development
}
\examples{
model_Bede()
}
\keyword{Model}
