% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_Bede.R
\name{plot_model_Bede}
\alias{plot_model_Bede}
\title{Model Plot Function}
\usage{
plot_model_Bede(
  df_mean,
  df_pred,
  Metric,
  Formula,
  xvar,
  colourvar,
  pallette,
  xlabel
)
}
\arguments{
\item{plot}{used to create ggplot scales}
}
\description{
This Function creates a ggplot of Model Predictions using a X variable (xvar) and treatment (colourvar)
}
\examples{
plot_model_Bede()
}
\keyword{Model}
\keyword{Predictions}
