% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Bede_MDS.R
\name{theme_Bede_MDS}
\alias{theme_Bede_MDS}
\title{A ggplot Theme Function}
\usage{
theme_Bede_MDS(...)
}
\description{
This function adds a custom theme to ggplots specifically for plotting AXIS less MDS plots
}
\examples{
df<- data.frame(x=rep(seq(1,10,1),10),y=rep(rnorm(10,4),10))
ggplot(df,aes(x=x,y=y))+geom_point()+theme_Bede_MDS()
}
\keyword{Theme}
